SCSS
I used Bibi's idea of using an already made website from the Codeacademy (she sent about it on slack). It really helped with time and less confusion.
I really like SCSS but I need more time to learn it as well as I'm still confused with CSS.
I hope I managed to do the minimum and that you like this random website.